requirements.txt
tensorflow
scikit-learn
numpy
pandas
matplotlib
seaborn
streamlit
pyautogui
pillow
opencv-python